// src/types/api.ts
export interface KnowledgeBase {
  id: number;
  name: string;
  description: string;
  knowledge_type: string;
  path: string;
  unique_code: string;
  knowledgebase_id: string;
  vector_table_name: string;
  chunk_size?: number;
  chunk_overlap?: number;
  created_at: string;
  updated_at: string;
  model_provider: number;
  is_file_linked?: boolean;
  permissions?: KnowledgeBasePermission[];
  permissions_input?: KnowledgeBasePermission[];
  role?: "owner" | "editor" | "viewer"
}
// ...
export interface Agent {
  id: number;
  instructions: Instruction;
  expected_output: ExpectedOutput;
  name: string;
  description: string;
  unique_code: string;
  agent_id: string; 
  session_table: string;
  search_knowledge: boolean;
  add_datetime_to_instructions: boolean;
  show_tool_calls: boolean;
  markdown_enabled: boolean;
  debug_mode: boolean;
  num_history_responses: number;
  is_global: boolean;
  created_at: string;
  updated_at: string;
  user: number;
  model: number;
  knowledge_base: string;
  team: number | null;
  subscriptions: number[];
}

export interface AgentCreate {
  name: string;
  description: string;
  category: number;
  model: number;
  instructions_id: number;
  custom_instruction: string;
  expected_output_id: number;
  custom_excpected_output: string;
  expected_output_data: {
    title: string;
    expected_output: string;
    category: string;
    is_enabled: boolean;
    is_global: boolean;
    user: number;
    agent: number;
  };
  knowledge_base: string;
  search_knowledge: boolean;
  cite_knowledge: boolean;
  add_datetime_to_instructions: boolean;
  show_tool_calls: boolean;
  markdown_enabled: boolean;
  debug_mode: boolean;
  num_history_responses: number;
  is_global: boolean;
  team: number | null;
  subscriptions: number[];
}
// ...
export interface Instruction {
  id: number;
  updated_at: string;
  title: string;
  instruction: string;
  category: string;
  is_template: boolean;
  is_enabled: boolean;
  is_global: boolean;
  is_system: boolean;
  created_at: string;
  user: number;
}

export interface ExpectedOutput {
  id: number;
  updated_at: string;
  title: string;
  expected_output: string;
  category: string;
  is_enabled: boolean;
  is_global: boolean;
  created_at: string;
  user: number;
  agent: number | null;
}